WPB-19085 merge conversation, subconversation and member stores (#4740)
