+++ libs/wire-subsystems/src/Wire/ConversationStore.hs
@@ -19,6 +19,7 @@
 
 module Wire.ConversationStore where
 
+import Data.Domain
 import Data.Id
 import Data.Misc
 import Data.Qualified
@@ -29,9 +30,16 @@ import Polysemy
 import Wire.API.Conversation hiding (Conversation, Member)
 import Wire.API.Conversation.CellsState
 import Wire.API.Conversation.Protocol
-import Wire.API.MLS.CipherSuite (CipherSuiteTag)
+import Wire.API.Conversation.Role
+import Wire.API.MLS.CipherSuite
+import Wire.API.MLS.Credential
 import Wire.API.MLS.GroupInfo
+import Wire.API.MLS.LeafNode
+import Wire.API.MLS.SubConversation
+import Wire.API.Provider.Service
+import Wire.ConversationStore.MLS.Types
 import Wire.StoredConversation
+import Wire.UserList
 
 data LockAcquired
   = Acquired
@@ -75,8 +83,45 @@ data ConversationStore m a where
   GetTeamConversation :: TeamId -> ConvId -> ConversationStore m (Maybe ConvId)
   GetTeamConversations :: TeamId -> ConversationStore m [ConvId]
   DeleteTeamConversations :: TeamId -> ConversationStore m ()
+  -- MEMBER OPERATIONS
+  CreateMembers :: (ToUserRole u) => ConvId -> UserList u -> ConversationStore m ([LocalMember], [RemoteMember])
+  CreateMembersInRemoteConversation :: Remote ConvId -> [UserId] -> ConversationStore m ()
+  CreateBotMember :: ServiceRef -> BotId -> ConvId -> ConversationStore m BotMember
+  GetLocalMember :: ConvId -> UserId -> ConversationStore m (Maybe LocalMember)
+  GetLocalMembers :: ConvId -> ConversationStore m [LocalMember]
+  GetAllLocalMembers :: ConversationStore m [LocalMember]
+  GetRemoteMember :: ConvId -> Remote UserId -> ConversationStore m (Maybe RemoteMember)
+  GetRemoteMembers :: ConvId -> ConversationStore m [RemoteMember]
+  CheckLocalMemberRemoteConv :: UserId -> Remote ConvId -> ConversationStore m Bool
+  SelectRemoteMembers :: [UserId] -> Remote ConvId -> ConversationStore m ([UserId], Bool)
+  SetSelfMember :: Qualified ConvId -> Local UserId -> MemberUpdate -> ConversationStore m ()
+  SetOtherMember :: Local ConvId -> Qualified UserId -> OtherMemberUpdate -> ConversationStore m ()
+  DeleteMembers :: ConvId -> UserList UserId -> ConversationStore m ()
+  DeleteMembersInRemoteConversation :: Remote ConvId -> [UserId] -> ConversationStore m ()
+  AddMLSClients :: GroupId -> Qualified UserId -> Set (ClientId, LeafIndex) -> ConversationStore m ()
+  PlanClientRemoval :: (Foldable f) => GroupId -> f ClientIdentity -> ConversationStore m ()
+  RemoveMLSClients :: GroupId -> Qualified UserId -> Set ClientId -> ConversationStore m ()
+  RemoveAllMLSClients :: GroupId -> ConversationStore m ()
+  LookupMLSClients :: GroupId -> ConversationStore m (ClientMap LeafIndex)
+  LookupMLSClientLeafIndices :: GroupId -> ConversationStore m (ClientMap LeafIndex, IndexMap)
+  GetRemoteMembersByDomain :: Domain -> ConversationStore m [(ConvId, RemoteMember)]
+  GetLocalMembersByDomain :: Domain -> ConversationStore m [(ConvId, UserId)]
+  -- SUB CONVERSATION OPERATIONS
+  CreateSubConversation :: ConvId -> SubConvId -> GroupId -> ConversationStore m SubConversation
+  GetSubConversation :: ConvId -> SubConvId -> ConversationStore m (Maybe SubConversation)
+  GetSubConversationGroupInfo :: ConvId -> SubConvId -> ConversationStore m (Maybe GroupInfoData)
+  GetSubConversationEpoch :: ConvId -> SubConvId -> ConversationStore m (Maybe Epoch)
+  SetSubConversationGroupInfo :: ConvId -> SubConvId -> Maybe GroupInfoData -> ConversationStore m ()
+  SetSubConversationEpoch :: ConvId -> SubConvId -> Epoch -> ConversationStore m ()
+  SetSubConversationCipherSuite :: ConvId -> SubConvId -> CipherSuiteTag -> ConversationStore m ()
+  ListSubConversations :: ConvId -> ConversationStore m (Map SubConvId ConversationMLSData)
+  DeleteSubConversation :: ConvId -> SubConvId -> ConversationStore m ()
 
 makeSem ''ConversationStore
 
 acceptConnectConversation :: (Member ConversationStore r) => ConvId -> Sem r ()
 acceptConnectConversation cid = setConversationType cid One2OneConv
+
+-- | Add a member to a local conversation, as an admin.
+createMember :: (Member ConversationStore r) => Local ConvId -> Local UserId -> Sem r [LocalMember]
+createMember c u = fst <$> createMembers (tUnqualified c) (UserList [tUnqualified u] [])
