+++ services/galley/src/Galley/API/Action/Reset.hs
@@ -31,7 +31,6 @@ import Wire.API.MLS.SubConversation
 import Wire.API.Routes.Public.Galley.MLS
 import Wire.API.VersionInfo
 import Wire.ConversationStore
-import Wire.MemberStore
 import Wire.NotificationSubsystem
 import Wire.Sem.Now (Now)
 import Wire.StoredConversation as Data
@@ -50,8 +49,6 @@ resetLocalMLSMainConversation ::
     Member Random r,
     Member Resource r,
     Member ConversationStore r,
-    Member MemberStore r,
-    Member SubConversationStore r,
     Member P.TinyLog r
   ) =>
   Qualified UserId ->
