+++ services/galley/src/Galley/API/Clients.hs
@@ -77,11 +77,9 @@ rmClient ::
     Member Now r,
     Member (ListItems p1 ConvId) r,
     Member (ListItems p1 (Remote ConvId)) r,
-    Member MemberStore r,
     Member (Error InternalError) r,
     Member ProposalStore r,
     Member Random r,
-    Member SubConversationStore r,
     Member P.TinyLog r,
     Member TeamCollaboratorsSubsystem r
   ) =>
