+++ services/galley/src/Galley/API/Create.hs
@@ -84,7 +84,6 @@ import Wire.API.Team.Permission hiding (self)
 import Wire.API.User
 import Wire.BrigAPIAccess
 import Wire.ConversationStore qualified as E
-import Wire.MemberStore qualified as E
 import Wire.NotificationSubsystem
 import Wire.Sem.Now (Now)
 import Wire.Sem.Now qualified as Now
@@ -124,8 +123,7 @@ createGroupConversationUpToV3 ::
     Member TeamStore r,
     Member P.TinyLog r,
     Member TeamFeatureStore r,
-    Member TeamCollaboratorsSubsystem r,
-    Member MemberStore r
+    Member TeamCollaboratorsSubsystem r
   ) =>
   Local UserId ->
   Maybe ConnId ->
@@ -170,8 +168,7 @@ createGroupOwnConversation ::
     Member TeamStore r,
     Member P.TinyLog r,
     Member TeamFeatureStore r,
-    Member TeamCollaboratorsSubsystem r,
-    Member MemberStore r
+    Member TeamCollaboratorsSubsystem r
   ) =>
   Local UserId ->
   Maybe ConnId ->
@@ -216,8 +213,7 @@ createGroupConversation ::
     Member TeamStore r,
     Member P.TinyLog r,
     Member TeamFeatureStore r,
-    Member TeamCollaboratorsSubsystem r,
-    Member MemberStore r
+    Member TeamCollaboratorsSubsystem r
   ) =>
   Local UserId ->
   Maybe ConnId ->
@@ -263,8 +259,7 @@ createGroupConvAndMkResponse ::
     Member LegalHoldStore r,
     Member TeamStore r,
     Member TeamFeatureStore r,
-    Member TeamCollaboratorsSubsystem r,
-    Member MemberStore r
+    Member TeamCollaboratorsSubsystem r
   ) =>
   Local UserId ->
   Maybe ConnId ->
@@ -305,8 +300,7 @@ createGroupConversationGeneric ::
     Member TeamStore r,
     Member P.TinyLog r,
     Member TeamFeatureStore r,
-    Member TeamCollaboratorsSubsystem r,
-    Member MemberStore r
+    Member TeamCollaboratorsSubsystem r
   ) =>
   Local UserId ->
   Maybe ConnId ->
@@ -704,7 +698,6 @@ createConnectConversation ::
     Member FederatorAccess r,
     Member NotificationSubsystem r,
     Member Now r,
-    Member MemberStore r,
     Member P.TinyLog r
   ) =>
   Local UserId ->
@@ -803,7 +796,7 @@ newRegularConversation ::
     Member (ErrorS OperationDenied) r,
     Member (Error InvalidInput) r,
     Member (Input Opts) r,
-    Member MemberStore r
+    Member ConversationStore r
   ) =>
   Local UserId ->
   NewConv ->
