+++ services/galley/src/Galley/API/LegalHold.hs
@@ -165,11 +165,9 @@ removeSettingsInternalPaging ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamFeatureStore r,
     Member (TeamMemberStore InternalPaging) r,
     Member TeamStore r,
@@ -212,11 +210,9 @@ removeSettings ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member (Embed IO) r,
     Member TeamCollaboratorsSubsystem r
   ) =>
@@ -267,13 +263,11 @@ removeSettings' ::
     Member (Input Env) r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member (TeamMemberStore p) r,
     Member TeamStore r,
     Member ProposalStore r,
     Member Random r,
     Member P.TinyLog r,
-    Member SubConversationStore r,
     Member (Embed IO) r,
     Member TeamCollaboratorsSubsystem r
   ) =>
@@ -319,11 +313,9 @@ grantConsent ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamStore r,
     Member TeamCollaboratorsSubsystem r
   ) =>
@@ -369,11 +361,9 @@ requestDevice ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamFeatureStore r,
     Member TeamStore r,
     Member (Embed IO) r,
@@ -464,11 +454,9 @@ approveDevice ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamFeatureStore r,
     Member TeamStore r,
     Member (Embed IO) r,
@@ -543,11 +531,9 @@ disableForUser ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamStore r,
     Member (Embed IO) r,
     Member TeamCollaboratorsSubsystem r
@@ -609,12 +595,10 @@ changeLegalholdStatusAndHandlePolicyConflicts ::
     Member Now r,
     Member LegalHoldStore r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member TeamStore r,
     Member ProposalStore r,
     Member Random r,
     Member P.TinyLog r,
-    Member SubConversationStore r,
     Member TeamCollaboratorsSubsystem r
   ) =>
   TeamId ->
@@ -726,11 +710,9 @@ handleGroupConvPolicyConflicts ::
     Member (Input Env) r,
     Member Now r,
     Member (ListItems LegacyPaging ConvId) r,
-    Member MemberStore r,
     Member ProposalStore r,
     Member P.TinyLog r,
     Member Random r,
-    Member SubConversationStore r,
     Member TeamStore r,
     Member TeamCollaboratorsSubsystem r
   ) =>
