+++ services/galley/src/Galley/API/Teams/Features.hs
@@ -330,9 +330,7 @@ instance SetFeatureConfig LegalholdConfig where
         Member Now r,
         Member LegalHoldStore r,
         Member (ListItems LegacyPaging ConvId) r,
-        Member MemberStore r,
         Member ProposalStore r,
-        Member SubConversationStore r,
         Member TeamFeatureStore r,
         Member TeamStore r,
         Member (TeamMemberStore InternalPaging) r,
